fiw warnings by compare as signed int

+++ include/qoa.h
@@ -366,7 +366,7 @@ unsigned int qoa_encode_frame(const short *sample_data, qoa_desc *qoa, unsigned 
     ), bytes, &p);
 
 
-    for (int c = 0; c < channels; c++) {
+    for (int c = 0; c < (int)channels; c++) {
         /* If the weights have grown too large, reset them to 0. This may happen
         with certain high-frequency sounds. This is a last resort and will
         introduce quite a bit of noise, but should at least prevent pops/clicks */
@@ -395,9 +395,9 @@ unsigned int qoa_encode_frame(const short *sample_data, qoa_desc *qoa, unsigned 
 
     /* We encode all samples with the channels interleaved on a slice level.
     E.g. for stereo: (ch-0, slice 0), (ch 1, slice 0), (ch 0, slice 1), ...*/
-    for (int sample_index = 0; sample_index < frame_len; sample_index += QOA_SLICE_LEN) {
+    for (int sample_index = 0; sample_index < (int)frame_len; sample_index += QOA_SLICE_LEN) {
 
-        for (int c = 0; c < channels; c++) {
+        for (int c = 0; c < (int)channels; c++) {
             int slice_len = qoa_clamp(QOA_SLICE_LEN, 0, frame_len - sample_index);
             int slice_start = sample_index * channels + c;
             int slice_end = (sample_index + slice_len) * channels + c;
@@ -490,7 +490,7 @@ void *qoa_encode(const short *sample_data, qoa_desc *qoa, unsigned int *out_len)
 
     unsigned char *bytes = (unsigned char *)QOA_MALLOC(encoded_size);
 
-    for (int c = 0; c < qoa->channels; c++) {
+    for (int c = 0; c < (int)qoa->channels; c++) {
         /* Set the initial LMS weights to {0, 0, -1, 2}. This helps with the
         prediction of the first few ms of a file. */
         qoa->lms[c].weights[0] = 0;
@@ -513,7 +513,7 @@ void *qoa_encode(const short *sample_data, qoa_desc *qoa, unsigned int *out_len)
     #endif
 
     int frame_len = QOA_FRAME_LEN;
-    for (int sample_index = 0; sample_index < qoa->samples; sample_index += frame_len) {
+    for (int sample_index = 0; sample_index < (int)qoa->samples; sample_index += frame_len) {
         frame_len = qoa_clamp(QOA_FRAME_LEN, 0, qoa->samples - sample_index);
         const short *frame_samples = sample_data + sample_index * qoa->channels;
         unsigned int frame_size = qoa_encode_frame(frame_samples, qoa, frame_len, bytes + p);
@@ -588,7 +588,7 @@ unsigned int qoa_decode_frame(const unsigned char *bytes, unsigned int size, qoa
     if (
         channels != qoa->channels ||
         samplerate != qoa->samplerate ||
-        frame_size > size ||
+        frame_size > (int)size ||
         samples * channels > max_total_samples
     ) {
         return 0;
